Keep current extent content in memory
This helps with avoid fetching the same object over and over when
consumers do reads shorter than the extent, for easy 200..300x speed
on sequential read over wifi.

At some point, a proper object cache is probably going to be needed,
but that's trickier to balance with regard to memory consumption. This
one is pretty obvious: each Reader (that actually has Read called, not
just ReadAt) holds up to one extent in memory (<= 16 MB, with current
tunables).

Ideally, we can keep the object cache outside, as just a
gocloud.dev/blob backend.

In that world, we should probably prepopulate the kernel page cache
with the current extent contents (see FUSE NotifyStore), and not cache
in userspace.