Add config field symlink_target, used by `plop add` for symlinks
`symlink_target` defaults to `mountpoint`, so the following is all
opt-in.

When `plop add` creates a symlink, it needs to point the target to a
plopfs mountpoint. Unix legacy means non-root users are not in control
of the filesystem namespace they see (or they could change the
effective /etc/passwd and such). This means per-user filesystems are
normally mounted only in subtrees that the user controls (/home/jdoe,
/run/user/1000, etc). With symlinks referring to these locations,
details about the user get stored in the symlink, and thus e.g. in the
git repository managing the symlinks. This will not work if these
trees are shared between multiple users.

(Would be nice if unix paths could refer to the current user's home.
`~` is a shell-level path rewriting rule and does not work here.)

To prepare for multi-user sharing, separate the notion of path where
the FUSE filesystem is mounted from the notion of symlink target. This
way, symlinks can refer to a global path (e.g. `/plop`) while the
filesystem is mounted in a user-controlled path (e.g.
`/home/jdoe/media/plop`). On a single-user host, the former can be
just a symlink to the latter (`/plop` -> `/home/jdoe/media/plop`),
while on a multi-user host something more complex is needed (a small
FUSE filesystem will suffice).

With this, symlink targets can be made futureproof. (Assuming that all
users are willing and able to have such customization in their file
tree, at the same location. There's limits to what unix allows.)

(An alternative design would be to serve all users with a single FUSE
filesystem instance. Since it deals with secrets, this is considered
unnecessarily risky and obtaining the right secrets would be awkward
to plumb. Nothing here prevents that future.)