cas: Cache extent contents
Previously the only cache was current extent for Read at its singular
offset. That didn't help ReadAt at all, and ReadAt is what FUSE needs.
This should help plopfs performance, at the cost of some RAM usage
(currently tuned to around ~300 MB per volume, not configurable at
this time).

Caching extents is important because we don't want to decrypt etc
an 8 MB extent, use only 4 kB of the result, and discard it -- only
to redo the effort for the next 4kB slice.