cas: Expose Reader.ExtentAt
I may regret the exact API (what returns an error and what doesn't),
but since there's no stability promise at this time that's fine for
now.

This is needed for debug etc tooling to be able to enumerate extents.
An alternative would be shoving all of this code into
bazil.org/plop/internal/ and then re-exposing parts of it in the cas
package; that's way too much bureaucracy.